+++ common/src/main/java/de/cech12/solarcooker/ModTags.java
@@ -2,6 +2,7 @@ package de.cech12.solarcooker;
 
 import net.minecraft.core.registries.Registries;
 import net.minecraft.tags.TagKey;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
 
 import javax.annotation.Nonnull;
@@ -14,6 +15,15 @@ public class ModTags {
 
         private static TagKey<Block> tag(@Nonnull String name) {
             return TagKey.create(Registries.BLOCK, Constants.id(name));
+        }
+    }
+
+    public static class Items {
+
+        public static final TagKey<Item> SOLAR_COOKER_REFLECTOR = tag("solar_cooker_reflector");
+
+        private static TagKey<Item> tag(@Nonnull String name) {
+            return TagKey.create(Registries.ITEM, Constants.id(name));
         }
     }
 }