+++ common/src/main/java/de/cech12/solarcooker/item/ReflectorItem.java
@@ -1,0 +1,23 @@
+package de.cech12.solarcooker.item;
+
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
+
+import javax.annotation.Nonnull;
+import java.util.List;
+
+public class ReflectorItem extends Item {
+
+    public ReflectorItem(Properties properties) {
+        super(properties);
+    }
+
+    @Override
+    public void appendHoverText(@Nonnull ItemStack stack, @Nonnull TooltipContext context, @Nonnull List<Component> tooltip, @Nonnull TooltipFlag flag) {
+        super.appendHoverText(stack, context, tooltip, flag);
+        tooltip.add(Component.translatable("item.solarcooker.reflector.description").withStyle(ChatFormatting.BLUE));
+    }
+}