+++ CHANGELOG.md
@@ -3,14 +3,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Forge Recommended Versioning](https://mcforge.readthedocs.io/en/latest/conventions/versioning/).
 
-## [1.21-4.0.0.0] - 2024-06-30
+## [1.21-4.0.0.0] - 2024-07-07
 ### Changed
 - updated to Minecraft 1.21 (Fabric 0.100.3+1.21, Neoforge 21.0.42-beta, Forge 51.0.18)
 - updated Cloth Config support (15.0.127) (Fabric/Quilt)
 - updated ModMenu support (11.0.1) (Fabric/Quilt)
-- updated JEI support (19.0.0.11) (Forge & NeoForge)
+- updated JEI support (19.0.0.11) (all loaders)
 - updated REI support (16.0.729) (Fabric/Quilt & NeoForge)
 - updated The One Probe support (1.21_neo-12.0.0) (NeoForge)
+- reflectors can now be placed in the Solar Cooker GUI and are rendered as additional lids
+- default reflector speed factor was raised to 0.5 (former value was 0.25) (configurable)
+- default recipe type of the Solar Cooker was changed to "smelting" (former value was "smoking") (configurable)
+
+### Removed
+- reflector blocks were removed (they will disappear when you update your world from an older version) 
 
 ## [1.20.6-3.2.0.0] - 2024-06-30
 ### Changed